requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
numpy==1.26.1
python-jose[cryptography]==3.3.0
pytest==7.4.3
httpx==0.25.1
redis==5.0.1
black==23.10.1
flake8==6.1.0